Refactors EntityManager to not do any networking. ServerEntityManager and ClientEntityManager now do the networking instead.
