+++ Robust.UnitTesting/Server/RobustServerSimulation.cs
@@ -175,8 +175,8 @@ namespace Robust.UnitTesting.Server
             container.RegisterInstance<IGameTiming>(new Mock<IGameTiming>().Object); // TODO: get timing working similar to RobustIntegrationTest
 
             //Tier 2: Simulation
-            container.Register<IServerEntityManager, ServerEntityManager>();
-            container.Register<IEntityManager, ServerEntityManager>();
+            //container.Register<IServerEntityManager, EntityManager>();
+            container.Register<IEntityManager, EntityManager>();
             container.Register<IMapManager, MapManager>();
             container.Register<IEntityLookup, SharedEntityLookup>();
             container.Register<ISerializationManager, SerializationManager>();
@@ -190,7 +190,7 @@ namespace Robust.UnitTesting.Server
 
             //Tier 3: Networking
             //TODO: Try to remove these
-            container.RegisterInstance<IEntityNetworkManager>(new Mock<IEntityNetworkManager>().Object);
+            //container.RegisterInstance<IEntityNetworkManager>(new Mock<IEntityNetworkManager>().Object);
             container.RegisterInstance<INetManager>(new Mock<INetManager>().Object);
 
             _diFactory?.Invoke(container);
